% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{download_exercise}
\alias{download_exercise}
\title{Download a exercise R Markdown file}
\usage{
download_exercise(exercise_number)
}
\arguments{
\item{exercise_number}{A number indicating which exercise number file to get.}
}
\description{
This function downloads exercise files from the class GitHub repository.
}
\examples{
 # Download the first exercise file
 \dontrun{download_exercise(1)}

}
