% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tki_demo}
\alias{tki_demo}
\title{tki_demo}
\format{A data table with 100 rows and 8 columns:
\describe{
  \item{\code{id}}{integer unique participant identifier}
  \item{\code{dob}}{dat date of birth}
  \item{\code{male}}{logical T - male; F - female}
  \item{\code{smoker}}{logical T - smoker; F - non-smoker}
  \item{\code{intervention}}{factor one of: Placebo, Drug 1, or Drug 2}
  \item{\code{day1}}{numeric measurement on day 1}
  \item{\code{day2}}{numeric measurement on day 2}
  \item{\code{day3}}{numeric measurement on day 3}
}}
\usage{
data(tki_demo)
}
\description{
Demonstration data set for the Telethon Kids Institute Introduction to R 
Workshop.
}
\keyword{datasets}
