% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tki_demo_complications}
\alias{tki_demo_complications}
\title{tki_demo_complications}
\format{A data table with 20 rows and 2 columns:
\describe{
  \item{\code{id}}{integer unique participant identifier}
  \item{\code{complications}}{text health complications that were experienced by the participant}
}}
\usage{
data(tki_demo_complications)
}
\description{
Demonstration data set for the Telethon Kids Institute Introduction to R 
Workshop. This second data frame is to illustrate joining data.
}
\keyword{datasets}
